back/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crontab==1.0.1
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
greenlet==3.0.3
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
playwright==1.43.0
prompt-toolkit==3.0.43
pycparser==2.21
pyee==11.1.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
